requirements.txt
# Core dependencies
streamlit==1.40.0
requests==2.32.3
python-dotenv==1.0.1
openai==1.56.2

# Data processing
pandas==2.2.3
numpy==1.26.4

# Date and time
python-dateutil==2.9.0

# XML parsing
lxml==5.3.0

# JSON handling
jsonschema==4.23.0

# Caching
cachetools==5.5.0

# Session management
streamlit-extras==0.4.7

# Charts and visualization
plotly==5.24.1
altair==5.4.1

# Document processing
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5

# Natural Language Processing (Korean)
konlpy==0.6.0
# Note: konlpy requires Java runtime

# Async support
aiohttp==3.11.10
asyncio==3.4.3

# Logging
loguru==0.7.2

# Testing (optional)
pytest==8.3.4
pytest-asyncio==0.25.0